requirements.txt
pyrr
pyopengl
glfw